chore: remove bugs from timer code
