--- src/components/Time/IntervalComponent/IntervalComponent.scss
@@ -1,54 +1,0 @@
-.interval-container {
-    margin-left: auto;
-    margin-right: auto;
-    // width: max-content;
-    text-align: center;
-}
-
-.interval-header-row {
-    font-size: $font-size-small;
-    margin-bottom: 1px;
-    text-align: center;
-}
-
-.text-center {
-    font-size: $font-size-large;
-    margin-bottom: 1px;
-}
-
-.buttonDefaults {
-    border-radius: 500px;
-    height: 60px;
-    width: 60px;
-}
-
-.startButton {
-    background: #39B0E6;
-    color: #fff;
-}
-
-.stopButton {
-    background: #D75452;
-}
-
-.resetButton {
-    background: #39B0E6;
-    color: #fff;
-}
-
-.text-primary {
-    color: #39B0E6;
-    text-align: left;
-    margin: 0 0 1px 10px;
-}
-
-.time-row {
-    display: flex;
-    justify-content: space-between;
-    margin: 10px 30px;
-    
-}
-
-.center-block {
-
-}
