+++ src/components/Time/TimerComponent/OldTimerComponent.scss
@@ -1,0 +1,146 @@
+// Basic reset
+
+
+.NumInput {
+  width: 260px;
+  height: auto;
+
+  margin: 30px auto;
+  padding: 20px 20px 9px;
+
+  border-radius: 3px;
+
+  /* Using box shadows to create 3D effects */
+  background: #9dd2ea;
+  background: linear-gradient(#9dd2ea, #8bceec);
+  border-radius: 3px;
+  box-shadow: 0px 4px #009de4, 0px 10px 15px rgba(0, 0, 0, 0.2);
+}
+
+/* Top portion */
+.top .timer-key.clear {
+  float: left;
+}
+
+/* Inset shadow on the screen to create indent */
+.top .screen {
+  height: 40px;
+  width: 212px;
+
+  float: right;
+
+  padding: 0 10px;
+
+  background: rgba(0, 0, 0, 0.2);
+  border-radius: 3px;
+  box-shadow: inset 0px 4px rgba(0, 0, 0, 0.2);
+
+  /* Typography */
+  font-size: 17px;
+  line-height: 40px;
+  color: white;
+  text-shadow: 1px 1px 2px rgba(0, 0, 0, 0.2);
+  text-align: right;
+  letter-spacing: 1px;
+}
+
+/* Clear floats */
+.keys {
+  overflow: hidden;
+}
+
+.keys .timer-key {
+  float: left;
+
+  width: 66px;
+  height: 40px;
+
+  margin: 0 7px 11px 0;
+}
+
+/* Applying same to the keys */
+.keys .timer-key,
+.top .timer-key.clear {
+  float: left;
+  position: relative;
+  top: 0;
+
+  cursor: pointer;
+
+  width: 66px;
+  height: 36px;
+
+  background: white;
+  border-radius: 3px;
+  box-shadow: 0px 4px rgba(0, 0, 0, 0.2);
+
+  margin: 0 7px 11px 0;
+
+  color: #888;
+  line-height: 36px;
+  text-align: center;
+
+  /* prevent selection of text inside keys */
+  user-select: none;
+
+  /* Smoothing out hover and active states using css3 transitions */
+  transition: all 0.2s ease;
+}
+
+/* Some hover effects */
+.keys .timer-key:hover {
+  background: $grey;
+  box-shadow: 0px 4px $blue;
+  color: white;
+}
+
+.keys .timer-key:active {
+  box-shadow: 0px 0px $blue;
+  top: 4px;
+}
+
+.empty-button {
+  float: left;
+  position: relative;
+  top: 0;
+
+  width: 66px;
+  height: 36px;
+
+  background: white;
+  border-radius: 3px;
+  box-shadow: 0px 4px rgba(0, 0, 0, 0.2);
+
+  margin: 0 7px 11px 0;
+
+  color: #888;
+  line-height: 36px;
+  text-align: center;
+
+
+
+}
+
+
+.timer-wrapper {
+  box-sizing: border-box;
+  padding: 0;
+  margin-left: auto;
+  margin-right: auto;
+  width: max-content;
+
+  /* Better text styling */
+  font: bold 14px Arial, sans-serif;
+}
+
+.timer-title {
+  font-size: $font-size-large;
+  margin-bottom: 1px;
+  text-align: center;
+}
+
+.play-buttons {
+  margin-left: auto;
+  margin-right: auto;
+  width: max-content;
+}
