+++ src/components/Time/TimerComponent/TimerConfig.js
