+++ src/styles/scss/styles.scss
@@ -11,7 +11,6 @@
 @import "../../components/Time/Time";
 @import "../../components/Time/TimerComponent/TimerComponent";
 @import "../../components/Time/StopwatchComponent/StopwatchComponent";
-@import "../../components/Time/IntervalComponent/IntervalComponent";
 @import "../../components/Badges/Badges";
 @import "../../components/BadgeCard/BadgeCard";
 @import "../../components/AddGoal/AddGoal";
